+++ packages/core/src/api/userActions/userAction.ts
@@ -6,7 +6,7 @@ import { ItemBuffer } from '../ItemBuffer';
 import { type MeasurementEvent } from '../measurements';
 import { type APIEvent } from '../types';
 
-import { UserActionSeverity } from './const';
+import { userActionEventName, UserActionSeverity } from './const';
 import { type HaltPredicate, type UserActionInterface, UserActionState } from './types';
 
 const defaultFollowUpActionTimeRange = 100;
@@ -164,8 +164,9 @@ export default class UserAction extends Observable implements UserActionInterfac
     this.notify(this._state);
 
     faro.api.pushEvent(
-      this.name,
+      userActionEventName,
       {
+        userActionName: this.name,
         userActionStartTime: this.startTime!.toString(),
         userActionEndTime: endTime.toString(),
         userActionDuration: duration.toString(),