+++ packages/core/src/api/userActions/initialize.ts
@@ -1,14 +1,16 @@
-import { type InternalLogger, type Transports } from '../..';
+import { type InternalLogger, type TransportItem, type Transports } from '../..';
 import type { Config } from '../../config';
 import { Observable } from '../../utils/reactive';
 
-import { UserActionSeverity, userActionStart, userActionStartByApiCallEventName } from './const';
+import { UserActionImportance, userActionStart, userActionStartByApiCallEventName } from './const';
 import {
   type StartUserActionOptions,
   type UserActionInterface,
+  type UserActionInternalInterface,
   type UserActionMessage,
   type UserActionsAPI,
   UserActionState,
+  type UserActionTransportItemBuffer,
 } from './types';
 import UserAction from './userAction';
 
@@ -44,7 +46,7 @@ export function initializeUserActionsAPI({
         transports,
         attributes,
         trigger: options?.triggerName || userActionStartByApiCallEventName,
-        severity: options?.severity || UserActionSeverity.Normal,
+        importance: options?.importance || UserActionImportance.Normal,
         trackUserActionsExcludeItem,
       });
       userAction
@@ -59,6 +61,7 @@ export function initializeUserActionsAPI({
         userAction: userAction,
       });
       activeUserAction = userAction;
+
       return activeUserAction;
     } else {
       internalLogger.error('Attempted to create a new user action while one is already running. This is not possible.');
@@ -70,8 +73,29 @@ export function initializeUserActionsAPI({
     return activeUserAction;
   };
 
-  return {
+  const api: UserActionsAPI = {
     startUserAction,
     getActiveUserAction,
   };
+
+  return api;
+}
+
+/**
+ * Adds an item to the buffer associated with the given UserAction.
+ * The item will only be added if the UserAction is in the Started state.
+ * @param userAction The UserAction instance
+ * @param item The item to add to the buffer
+ * @returns {boolean} True if the item was added, false otherwise
+ */
+export function addItemToUserActionBuffer(userAction: UserActionInterface | undefined, item: TransportItem): boolean {
+  if (!userAction) {
+    return false;
+  }
+  const state = (userAction as unknown as UserActionInternalInterface)?.getState();
+  if (state !== UserActionState.Started) {
+    return false;
+  }
+  (userAction as unknown as UserActionTransportItemBuffer).addItem(item);
+  return true;
 }