+++ packages/core/src/api/userActions/types.ts
@@ -1,6 +1,6 @@
 import { type TransportItem } from '../../transports';
 
-import { UserActionSeverity, userActionStartByApiCallEventName } from './const';
+import { type UserActionImportanceType, userActionStartByApiCallEventName } from './const';
 
 export enum UserActionState {
   Started,
@@ -9,18 +9,28 @@ export enum UserActionState {
   Ended,
 }
 
-export type HaltPredicate = () => boolean;
-
+/**
+ * Public interface for the UserAction.
+ * This is the interface that is part of the public API.
+ */
 export interface UserActionInterface {
   name: string;
   parentId: string;
+}
 
-  addItem(item: TransportItem): void;
-  extend(haltPredicate?: HaltPredicate): void;
+/**
+ * Internal interface for the UserAction.
+ * This interface is intended for internal use only and not guaranteed to be stable.
+ */
+export interface UserActionInternalInterface extends UserActionInterface {
+  halt(): void;
   end(attributes?: Record<string, string>): void;
-  halt(reason?: string): void;
   cancel(): void;
   getState(): UserActionState;
+}
+
+export interface UserActionTransportItemBuffer {
+  addItem(item: TransportItem): boolean;
 }
 
 export type ApiUserActionEvent = {
@@ -40,7 +50,7 @@ export type EndUserActionProps = {
 
 export type StartUserActionOptions = {
   triggerName?: string;
-  severity?: UserActionSeverity;
+  importance?: UserActionImportanceType;
 };
 
 export interface UserActionsAPI {