+++ packages/web-sdk/src/index.ts
@@ -98,6 +98,7 @@ export {
   Observable,
   unknownString,
   UserActionState,
+  type UserActionInternalInterface,
 } from '@grafana/faro-core';
 
 export type {
@@ -162,7 +163,7 @@ export type {
   TransportItemPayload,
   Transports,
   UnpatchedConsole,
-  UserActionSeverity,
+  UserActionImportance,
 } from '@grafana/faro-core';
 
 export {