+++ packages/web-sdk/src/instrumentations/userActions/processUserActionEventHandler.ts
@@ -1,22 +1,11 @@
-import { Observable, UserActionState } from '@grafana/faro-core';
-import type { Faro, Subscription, UserActionInterface } from '@grafana/faro-core';
+import type { Faro, Subscription, UserActionInterface, UserActionInternalInterface } from '@grafana/faro-core';
 
-import { monitorDomMutations, monitorHttpRequests, monitorPerformanceEntries } from '../_internal/monitors';
-import type {
-  HttpRequestEndMessage,
-  HttpRequestMessagePayload,
-  HttpRequestStartMessage,
-} from '../_internal/monitors/types';
-
-import { MESSAGE_TYPE_HTTP_REQUEST_END, MESSAGE_TYPE_HTTP_REQUEST_START, userActionDataAttribute } from './const';
+import { userActionDataAttributeParsed as userActionDataAttribute } from './const';
+import { UserActionController } from './userActionController';
 import { convertDataAttributeName } from './util';
 
 export function getUserEventHandler(faro: Faro) {
   const { api, config } = faro;
-
-  const httpMonitor = monitorHttpRequests();
-  const domMutationsMonitor = monitorDomMutations();
-  const performanceEntriesMonitor = monitorPerformanceEntries();
 
   function processUserEvent(event: PointerEvent | KeyboardEvent) {
     const userActionName = getUserActionNameFromElement(
@@ -31,55 +20,16 @@ export function getUserEventHandler(faro: Faro) {
 
     const userAction = api.startUserAction(userActionName, {}, { triggerName: event.type });
     if (userAction) {
-      proceessUserActionStarted(userAction);
+      processUserActionStarted(userAction);
     }
   }
 
-  function proceessUserActionStarted(userAction: UserActionInterface) {
-    const runningRequests = new Map<string, HttpRequestMessagePayload>();
-    const allMonitorsSub = new Observable()
-      .merge(httpMonitor, domMutationsMonitor, performanceEntriesMonitor)
-      .takeWhile(() => [UserActionState.Started, UserActionState.Halted].includes(userAction.getState()))
-      .filter((msg) => {
-        // If the user action is in halt state, we only keep listening to ended http requests
-        if (
-          userAction.getState() === UserActionState.Halted &&
-          !(isRequestEndMessage(msg) && runningRequests.has(msg.request.requestId))
-        ) {
-          return false;
-        }
-
-        return true;
-      })
-      .subscribe((msg) => {
-        if (isRequestStartMessage(msg)) {
-          // An action is on halt if it has pending items, like pending HTTP requests.
-          // In this case we start a separate timeout to wait for the requests to finish
-          // If in the halt state, we stop adding Faro signals to the action's buffer (see userActionLifecycleHandler.ts)
-          // But we are still subscribed to
-          runningRequests.set(msg.request.requestId, msg.request);
-        }
-
-        if (isRequestEndMessage(msg)) {
-          runningRequests.delete(msg.request.requestId);
-        }
-
-        if (!isRequestEndMessage(msg)) {
-          userAction.extend(() => runningRequests.size > 0);
-        } else if (userAction.getState() === UserActionState.Halted && runningRequests.size === 0) {
-          userAction.end();
-        }
-      });
-
-    (userAction as unknown as Observable)
-      .filter((v: UserActionState) => [UserActionState.Ended, UserActionState.Cancelled].includes(v))
-      .first()
-      .subscribe(() => {
-        unsubscribeAllMonitors(allMonitorsSub);
-      });
+  function processUserActionStarted(userAction: UserActionInterface) {
+    const internalUserAction = userAction as unknown as UserActionInternalInterface;
+    new UserActionController(internalUserAction).attach();
   }
 
-  return { processUserEvent, proceessUserActionStarted };
+  return { processUserEvent, processUserActionStarted };
 }
 
 export function getUserActionNameFromElement(element: HTMLElement, dataAttributeName: string): string | undefined {
@@ -98,12 +48,4 @@ export function getUserActionNameFromElement(element: HTMLElement, dataAttribute
 export function unsubscribeAllMonitors(allMonitorsSub: Subscription | undefined) {
   allMonitorsSub?.unsubscribe();
   allMonitorsSub = undefined;
-}
-
-export function isRequestStartMessage(msg: any): msg is HttpRequestStartMessage {
-  return msg.type === MESSAGE_TYPE_HTTP_REQUEST_START;
-}
-
-export function isRequestEndMessage(msg: any): msg is HttpRequestEndMessage {
-  return msg.type === MESSAGE_TYPE_HTTP_REQUEST_END;
 }