+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Fix (`@grafana/faro-web-sdk`): Fixed an issue where custom severity and custom trigger properties
   were not being included in user action attributes (#1551)
+- Fix (`@grafana/faro-web-sdk`): Fixed an error when `initializeFaro` is called without any window
+  object present (#1643)
 
 ## 2.0.0.beta-2